+++ src/lint/test/rules/index.ts
@@ -1,2 +1,1 @@
-export { createLintFixtureAdapter } from "./fixture-adapter.js";
-export { createLintFixtureSuiteDefinition } from "./fixture-suite-definition.js";
+export { createLintFixtureAdapter, createLintFixtureSuiteDefinition } from "./fixture-adapter.js";
