+++ src/refactor/test/fixture-adapter.ts
@@ -2,7 +2,7 @@ import { readFile, writeFile } from "node:fs/promises";
 import path from "node:path";
 
 import { Core } from "@gmloop/core";
-import type { FixtureAdapter } from "@gmloop/fixture-runner";
+import { type FixtureAdapter, FixtureRunner } from "@gmloop/fixture-runner";
 
 import { createPathSelectionMatcher } from "../src/codemods/naming-convention/path-selection.js";
 import { normalizeRefactorProjectConfig } from "../src/project-config.js";
@@ -176,5 +176,23 @@ export function createRefactorFixtureAdapter(): FixtureAdapter {
                 changed: true
             };
         }
+    });
+}
+
+/**
+ * Create the canonical refactor fixture suite definition shared by workspace
+ * and aggregate fixture runs.
+ *
+ * @returns Refactor fixture suite registration metadata.
+ */
+export function createRefactorFixtureSuiteDefinition() {
+    return FixtureRunner.createFixtureSuiteDefinition({
+        workspaceName: "refactor",
+        suiteName: "refactor fixtures",
+        compiledWorkspaceTestFilePath: "src/refactor/dist/test/refactor-fixtures.test.js",
+        moduleUrl: import.meta.url,
+        sourceRelativeSegments: ["fixtures"],
+        distRelativeSegments: ["..", "..", "test", "fixtures"],
+        adapter: createRefactorFixtureAdapter()
     });
 }
