+++ src/refactor/test/index.ts
@@ -1,2 +1,1 @@
-export { createRefactorFixtureAdapter } from "./fixture-adapter.js";
-export { createRefactorFixtureSuiteDefinition } from "./fixture-suite-definition.js";
+export { createRefactorFixtureAdapter, createRefactorFixtureSuiteDefinition } from "./fixture-adapter.js";
