+++ test/fixture-profile-adapter-coherence.test.ts
@@ -24,13 +24,13 @@ void test("fixture profile report uses the same workspace adapter factories as t
         readRepoFile("test/fixture-profile-report.ts"),
         readRepoFile("test/fixture-suite-registry.ts"),
         readRepoFile("src/format/test/formatter-fixtures.test.ts"),
-        readRepoFile("src/format/test/fixture-suite-definition.ts"),
+        readRepoFile("src/format/test/fixture-adapter.ts"),
         readRepoFile("src/format/src/format-entry.ts"),
         readRepoFile("src/lint/test/rules/rule-fixtures.test.ts"),
-        readRepoFile("src/lint/test/rules/fixture-suite-definition.ts"),
+        readRepoFile("src/lint/test/rules/fixture-adapter.ts"),
         readRepoFile("src/lint/src/index.ts"),
         readRepoFile("src/refactor/test/refactor-fixtures.test.ts"),
-        readRepoFile("src/refactor/test/fixture-suite-definition.ts"),
+        readRepoFile("src/refactor/test/fixture-adapter.ts"),
         readRepoFile("src/refactor/src/index.ts")
     ]);
 
@@ -42,10 +42,13 @@ void test("fixture profile report uses the same workspace adapter factories as t
 
     assert.match(formatSuiteSource, /createFormatFixtureSuiteDefinition\(\)/u);
     assert.match(formatDefinitionSource, /createFormatFixtureAdapter\(\)/u);
+    assert.match(formatDefinitionSource, /createFormatFixtureSuiteDefinition\(\)/u);
     assert.match(lintSuiteSource, /createLintFixtureSuiteDefinition\(\)/u);
     assert.match(lintDefinitionSource, /createLintFixtureAdapter\(\)/u);
+    assert.match(lintDefinitionSource, /createLintFixtureSuiteDefinition\(\)/u);
     assert.match(refactorSuiteSource, /createRefactorFixtureSuiteDefinition\(\)/u);
     assert.match(refactorDefinitionSource, /createRefactorFixtureAdapter\(\)/u);
+    assert.match(refactorDefinitionSource, /createRefactorFixtureSuiteDefinition\(\)/u);
     assert.doesNotMatch(formatRuntimeSource, /\btesting\b/u);
     assert.doesNotMatch(lintRuntimeSource, /\btesting\b/u);
     assert.doesNotMatch(refactorRuntimeSource, /\btesting\b/u);
